requirements.txt
accelerate==1.4.0.dev0
diffusers==0.33.0.dev0
einops==0.8.1
gradio==5.27.1
ipdb==0.13.13
matplotlib==3.10.1
MultiScaleDeformableAttention==1.0
numpy==2.2.5
openai==1.76.0
opencv_contrib_python==4.11.0.86
opencv_python==4.9.0.80
opencv_python_headless==4.11.0.86
Pillow==11.2.1
rembg==2.0.65
Requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
setuptools==60.2.0
timm==1.0.15
torch==2.4.0
torchsde==0.2.6
torchvision==0.19.0
tqdm==4.67.1
transformers==4.49.0
